ci: tighten up parse credit card rule
I have come across a sample that compare with all ASCII characters, which
triggers this rule (although the sample doesn't parse credit card
information). Ensure that there is not a comparison with `m`, `n`, `o`
and `p` (as they are in the middle of the alphabet and not part of the
hex char set) to avoid this case.