Rule for accessing Windows directory from KUSER_SHARED_DATA
This structure is at a fixed offset in both 32 and 64 bit Windows. This rule looks for a typical function call setup to copy data from this location.

It may be possible to instead look for only the known memory address, which would catch samples that manually access memory rather than calling a function. However, this would be more prone to false positives where the same constants are used for other reasons.